+++ template/src/{{ project_slug }}/frontend/oauth.py
@@ -7,8 +7,7 @@ from fastapi.responses import RedirectResponse
 from pydantic import BaseModel, Field
 from starlette.authentication import BaseUser
 
-from ..models.oauth import OAuthAccountModel
-from ..models.user import UserModel
+from ..models.core import OAuthAccountModel, UserModel
 from .routes import get_homepage_url
