+++ template/src/{{ project_slug }}/models/core/__init__.py
@@ -2,9 +2,16 @@ from beanie import Document, View
 
 from .blob import BlobModel
 from .email_verification import EmailVerificationTokenModel
+from .mixins import TimeStampMixin
 from .oauth import OAuthAccountModel
+from .types import Link
 from .user import UserModel
 
+
+__all__ = [
+    "TimeStampMixin",
+    "Link",
+]
 
 CORE_MODELS: list[type[Document] | type[View]] = [
     BlobModel,
