+++ template/src/{{ project_slug }}/models/core/mixins.py
@@ -5,7 +5,7 @@ from typing import Self
 from beanie import Insert, Replace, Save, SaveChanges, Update, before_event
 from pydantic import BaseModel, Field
 
-from ..time import Unit, now
+from ...time import Unit, now
 
 
 class Since(StrEnum):
