+++ template/src/{{ project_slug }}/models/core/oauth.py
