+++ template/src/{{ project_slug }}/mongo.py
@@ -2,8 +2,7 @@ from beanie import init_beanie
 from motor.motor_asyncio import AsyncIOMotorClient, AsyncIOMotorDatabase
 
 from ._config import project_settings, settings
-from .models.app import APP_MODELS
-from .models.core import CORE_MODELS
+from .models import MODELS
 
 
 # Include here any custom things that need to be done after model definitions
@@ -18,4 +17,4 @@ async def init_models() -> None:
         compressors=["zstd"],
     )[settings.mongo_dbname]
 
-    await init_beanie(database=mongo_db, document_models=CORE_MODELS + APP_MODELS)
+    await init_beanie(database=mongo_db, document_models=MODELS)
