--- template/src/{{ project_slug }}/mongo_types.py
@@ -1,9 +1,0 @@
-from typing import TYPE_CHECKING
-
-from beanie import Link as BeanieLink
-
-
-if TYPE_CHECKING:
-    type Link[T] = T | BeanieLink[T]
-else:
-    Link = BeanieLink
